fix coverage + reorg testing utils
