+++ tests/test_core/test_layout.py
@@ -181,10 +181,7 @@ async def test_layout_render_error_has_partial_update_with_error_message():
     def BadChild():
         raise ValueError("error from bad child")
 
-    with assert_idom_logged(
-        match_error="error from bad child",
-        clear_after=True,
-    ):
+    with assert_idom_logged(match_error="error from bad child"):
 
         with idom.Layout(Main()) as layout:
             patch = await render_json_patch(layout)
@@ -240,10 +237,7 @@ async def test_layout_render_error_has_partial_update_without_error_message():
     def BadChild():
         raise ValueError("error from bad child")
 
-    with assert_idom_logged(
-        match_error="error from bad child",
-        clear_after=True,
-    ):
+    with assert_idom_logged(match_error="error from bad child"):
 
         with idom.Layout(Main()) as layout:
             patch = await render_json_patch(layout)
@@ -743,7 +737,6 @@ async def test_duplicate_sibling_keys_causes_error(caplog):
         with assert_idom_logged(
             error_type=ValueError,
             match_error=r"Duplicate keys \['duplicate'\] at '/children/0'",
-            clear_after=True,
         ):
             await layout.render()
 
@@ -757,7 +750,6 @@ async def test_duplicate_sibling_keys_causes_error(caplog):
         with assert_idom_logged(
             error_type=ValueError,
             match_error=r"Duplicate keys \['duplicate'\] at '/children/0'",
-            clear_after=True,
         ):
             await layout.render()
 
@@ -796,10 +788,7 @@ async def test_log_error_on_bad_event_handler():
 
         return idom.html.button({"onClick": raise_error})
 
-    with assert_idom_logged(
-        match_error="bad event handler",
-        clear_after=True,
-    ):
+    with assert_idom_logged(match_error="bad event handler"):
 
         with idom.Layout(ComponentWithBadEventHandler()) as layout:
             await layout.render()
@@ -1233,7 +1222,6 @@ async def test_component_error_in_should_render_is_handled_gracefully():
         match_message=r".* component failed to check if .* should be rendered",
         error_type=ValueError,
         match_error="The error message",
-        clear_after=True,
     ):
         with idom.Layout(Root()) as layout:
             await layout.render()
